requirements.txt
coverage==4.5.1
Django==1.11
django-webpack-loader==0.6.0
djangorestframework==3.8.2
pytz==2018.4
